7.28

+++ astrum/Json/LoginUser.cs
@@ -19,8 +19,8 @@ namespace Astrum.Json
 
         public string name { get; set; }
 
-        public long minstaminastock { get; set; }
+        public int minstaminastock { get; set; }
 
-        public long minbpstock { get; set; }
+        public int minbpstock { get; set; }
     }
 }
